--- src/main/java/baseball/InputView.java
@@ -1,7 +1,0 @@
-package baseball;
-
-public class InputView {
-
-    // input 값을 입력받는 UI 클래스
-
-}
