+++ src/main/java/baseball/domain/Input.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 public class Input {
