+++ src/main/java/baseball/domain/Menu.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 public class Menu {
