+++ src/main/java/baseball/view/InputView.java
@@ -1,0 +1,13 @@
+package baseball.view;
+
+import java.util.Scanner;
+
+public final class InputView {
+
+    private static final Scanner scanner = new Scanner(System.in);
+
+    public String readInputNumber() {
+        return scanner.nextLine();
+    }
+
+}
